requirements.txt
boto
pynacl
